common: Add 'problem' property to CockpitPipe
This can be used to retrieve the problem that the pipe closed
with, or NULL. But more importantly, it can be used to construct
a CockpitPipe for a failure.

Since CockpitPipe guarantees that it won't fire it's close
signals until the main loop, we follow that same guarantee
here with the "problem" property on construct.