ws: Support multiple modules in a resource URL
Resource URLs now contain multiple modules, so that they can
contain base modules and derived modules.

They look like this:

/res/+module/+module@host/+xxchecksumyy/path/to/resource.file

Each part of the URL with a + is a module. After the + comes either
module or module@host or a checksum.

The resource is only looked up in the last module. The previous
modules are there so that the derived module can have ../ paths
in their code, and caching, lookups and everything work easily.