base: Move jquery bootstrap and term.js into new base module
And make the shell depend on this module.

The module is called 'latest' because it contains latest unstable
Cockpit API. There will be counterpart modules in the future.

Simplify the way the shell manifest works as well. This is a bit
non-standard since the shell is not loaded into its own frame/document.