base: Move spawn.js and channel.js into base package
So they can be used from multiple packages. We also document them
in the guide and clean up the API a bit.

One minor change is that channels now automatically detect which
host they should connect to based on the page parameters.

Another is that the 'expect disconnect' logic is handled at the
lowest level, so it is available to all components.

cockpit.logout() also becomes a low level function which
integrates well with the other transport logic.