kubernetes: Use 'kubectl proxy' to connect to kubernetes
This continues along the theme of Cockpit proving access to things
that would be otherwise accessible from the console.

This handles things like using the right IP address to connect to
openshift (even when local), using the right certificates, and
so on.

We can also remove a bunch of logic around parsing kubectl config
certificate and connection info in javascript.

Closes #2463