ws: Fix test-server deadlock when it starts
We tend to leave around dbus-daemon processes after our unit
tests are running. On certain operating systems like Fedora 23
this leads to a problem where eventually further dbus-daemon
processes cannot be launched.

Take over control of the launching of DBus daemon during tests.
Share code with the bridge to accomplish this.

In addition this gets rid of the icky fork() without exec() code
in g_test_dbus_up(). This is problematic due to the threading
behavior of GLib.

Closes #3215