bridge: Move the external bridge code along with shim
First of all we use a simple common function to look up the
channel type to instantiate.  This will let us determine the
channel type on things other than just the payload.

We also move the external bridge logic into the cockpitshim.c
file which brings it much closer to the code that uses it.
The few functions that are exported are only to be used by tests.